Make markdup QC optional, add samtools stats for duplicate metrics
