+++ cpg_workflows/stages/joint_genotyping.py
@@ -105,51 +105,3 @@ class JointGenotyping(CohortStage):
         for job in jobs:
             assert job
         return self.make_outputs(cohort, data=self.expected_outputs(cohort), jobs=jobs)
-
-
-@stage(required_stages=JointGenotyping)
-class JointVcfHappy(SampleStage):
-    """
-    Run Happy to validate validation samples in joint VCF
-    """
-
-    def expected_outputs(self, sample: Sample) -> Path | None:
-        """
-        Parsed by MultiQC: '*.summary.csv'
-        https://multiqc.info/docs/#hap.py
-        """
-        if sample.participant_id not in get_config().get('validation', {}).get(
-            'sample_map', {}
-        ):
-            return None
-
-        h = self.cohort.alignment_inputs_hash()
-        return (
-            self.cohort.analysis_dataset.prefix()
-            / 'qc'
-            / 'jc'
-            / 'hap.py'
-            / f'{h}-{sample.id}.summary.csv'
-        )
-
-    def queue_jobs(self, sample: Sample, inputs: StageInput) -> StageOutput | None:
-        """Queue jobs"""
-        vcf_path = inputs.as_path(target=self.cohort, stage=JointGenotyping, key='vcf')
-
-        jobs = happy(
-            b=self.b,
-            sample=sample,
-            vcf_or_gvcf=self.b.read_input_group(
-                **{
-                    'vcf.gz': str(vcf_path),
-                    'vcf.gz.tbi': str(vcf_path) + '.tbi',
-                }
-            ),
-            is_gvcf=False,
-            job_attrs=self.get_job_attrs(sample),
-            output_path=self.expected_outputs(sample),
-        )
-        if not jobs:
-            return self.make_outputs(sample)
-        else:
-            return self.make_outputs(sample, self.expected_outputs(sample), jobs)
