+++ main.py
@@ -12,7 +12,8 @@ from cpg_utils import to_path
 from cpg_utils.config import set_config_paths
 from cpg_workflows.workflow import run_workflow, StageDecorator
 from cpg_workflows.stages.large_cohort import LoadVqsr, Frequencies
-from cpg_workflows.stages.multiqc import GvcfMultiQC, CramMultiQC
+from cpg_workflows.stages.cram_qc import CramMultiQC
+from cpg_workflows.stages.gvcf_qc import GvcfMultiQC
 from cpg_workflows.stages.fastqc import FastQCMultiQC
 from cpg_workflows.stages.seqr_loader import MtToEs
 from cpg_workflows.stages.gatk_sv import GatherBatchEvidence
